break up main gql resolver too to make it more maintainable
